Fix build: ensure remaining is double in bookings list via clamp(0.0, total).toDouble(); align _refreshRoomOccupancy call in booking_edit to signature (remove stale previousRoom arg)

+++ mobile/lib/screens/bookings/bookings_list.dart
@@ -390,11 +390,11 @@ class _BookingRow extends ConsumerWidget {
       stream: paymentsRepo.paymentsByBooking(booking.id),
       builder: (context, snapshot) {
         final paid = snapshot.hasData ? snapshot.data!.fold<double>(0, (s, p) => s + p.amount) : 0.0;
-        final remaining = (totalAmount - paid).clamp(0, totalAmount);
-        final Color statusColor = remaining <= 0
+        final remaining = (totalAmount - paid).clamp(0.0, totalAmount).toDouble();
+        final Color statusColor = remaining <= 0.0
             ? Colors.green
             : (paid > 0 ? Colors.orange : Colors.red);
-        final String statusText = remaining <= 0
+        final String statusText = remaining <= 0.0
             ? 'مسددة'
             : (paid > 0 ? 'جزئياً' : 'غير مسددة');
